move _divide_dataset tests to test_trainer
