+++ tests/test_trainer.py
@@ -6,7 +6,8 @@ import pytest
 import logging
 import warnings
 import torch
-from deeprankcore.trainer import Trainer
+import h5py
+from deeprankcore.trainer import Trainer, _divide_dataset
 from deeprankcore.dataset import GraphDataset
 from deeprankcore.neuralnets.ginet import GINet
 from deeprankcore.neuralnets.foutnet import FoutNet
@@ -475,6 +476,44 @@ class TestTrainer(unittest.TestCase):
                     dataset_train = dataset_train,
                     dataset_test = dataset_test,
                     pretrained_model="test.pth.tar")
+
+    def test_trainsize(self):
+        hdf5 = "tests/data/hdf5/train.hdf5"
+        hdf5_file = h5py.File(hdf5, 'r')    # contains 44 datapoints
+        n_val = int ( 0.25 * len(hdf5_file) )
+        n_train = len(hdf5_file) - n_val
+        test_cases = [None, 0.25, n_val]
+        
+        for t in test_cases:
+            dataset_train, dataset_val =_divide_dataset(
+                dataset = GraphDataset(hdf5_path=hdf5),
+                splitsize=t,
+            )
+
+            assert len(dataset_train) == n_train
+            assert len(dataset_val) == n_val
+
+        hdf5_file.close()
+        
+    def test_invalid_trainsize(self):
+        hdf5 = "tests/data/hdf5/train.hdf5"
+        hdf5_file = h5py.File(hdf5, 'r')    # contains 44 datapoints
+        n = len(hdf5_file)
+        test_cases = [
+            1.0, n,     # cannot be 100% validation data
+            -0.5, -1,   # no negative values 
+            1.1, n + 1, # cannot use more than all data as input
+            ]
+        
+        for t in test_cases:
+            print(t)
+            with self.assertRaises(ValueError):
+                _divide_dataset(
+                    dataset = GraphDataset(hdf5_path=hdf5),
+                    splitsize=t,
+                )
+        
+        hdf5_file.close()
 
 
 if __name__ == "__main__":
